seperated into components
