+++ dreamscape-react-native/App.js
@@ -1,225 +1,17 @@
-import { setStatusBarHidden, StatusBar } from 'expo-status-bar';
-import { StyleSheet, View, FlatList } from 'react-native';
-import { Header as HeaderRNE, HeaderProps, Icon, ThemeProvider } from '@rneui/themed';
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, View} from 'react-native';
+import { Header as HeaderRNE } from '@rneui/themed';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import {Dimensions} from 'react-native';
-import { Image, Input, Button, Text} from '@rneui/themed';
-import React, { useState, useEffect, useMemo } from 'react';
-import { SearchOutlined } from '@ant-design/icons';
-import axios from 'axios';
-import {
-  MaterialReactTable,
-  useMaterialReactTable,
-} from 'material-react-table';
+import { Image, Text} from '@rneui/themed';
+import Register from './components/Register';
+import Search from './components/Search';
+import AllDataTable from './components/AllDataTable';
 
 const logo = require('./assets/logo.png')
 const windowWidth = Dimensions.get('window').width;
-const screenWidth = Dimensions.get('screen').width;
 
 export default function App() {
-  const [firstName, setFirstName] = useState('');
-  const [lastName, setLastName] = useState('');
-  const [dob, setDOB] = useState('');
-  const [gender, setGender] = useState('');
-
-  const [allData, setAllData] = useState([]);
-  const [deviceId, setDeviceId] = useState('');
-  const[searchLast, setSearchLast] = useState('');
-  const [searchResults, setSearchResults] = useState([]);
-
-  useEffect(() => {
-    axios
-      .get('http://localhost:8000/api/user/list')
-      .then((response) => {
-        // Handle success (e.g., display search results)
-        console.log('All patients:', response.data);
-        setAllData(response.data);
-        // You can handle the search results as needed for your application
-      })
-      .catch((error) => {
-        // Handle error
-        alert('Error retrieving patients. Please try again.');
-        if (error.response) {
-          console.error('Server responded with non-2xx status:', error.response.status);
-          console.error('Response data:', error.response.data);
-          console.error('Response headers:', error.response.headers);
-        } else if (error.request) {
-          // The request was made but no response was received
-          console.error('No response received from the server');
-          console.error('Request data:', error.request);
-        } else {
-          // Something happened in setting up the request that triggered an Error
-          console.error('Error setting up the request:', error.message);
-        }
-        console.error('Full error object:', error);
-      });
-  }, []);
-
-  const handleSearch = () => {
-    let apiUrl = '';
-    if (deviceId) {
-      apiUrl = `http://localhost:8000/api/user?device_id=${deviceId}`;
-    } else if (searchLast) {
-      apiUrl = `http://localhost:8000/api/user/search?last_name=${searchLast}`;
-    }
-
-    console.log('Searching for patient with:', deviceId ? 'Device ID ' + deviceId : 'Last Name ' + searchLast);
-    axios
-      .get(apiUrl)
-      .then((response) => {
-        // Handle success (e.g., display search results)
-        if (response.data === null || response.data.length === 0) {
-          // If response data is null or empty, display an alert
-          Alert.alert('No Results', 'No matching results found.');
-        } else {
-          // Update search results state
-          console.log('Search results:', response.data);
-          if (deviceId) {
-            setSearchResults(response.data);
-          }
-          else if (searchLast) {
-            setSearchResults(response.data[0]);
-          }
-        }
-        // You can handle the search results as needed for your application
-      })
-      .catch((error) => {
-        // Handle error
-        alert('No user found with that Device ID. Please try again.');
-        if (error.response) {
-          console.error('Server responded with non-2xx status:', error.response.status);
-          console.error('Response data:', error.response.data);
-          console.error('Response headers:', error.response.headers);
-        } else if (error.request) {
-          // The request was made but no response was received
-          console.error('No response received from the server');
-          console.error('Request data:', error.request);
-        } else {
-          // Something happened in setting up the request that triggered an Error
-          console.error('Error setting up the request:', error.message);
-        }
-        console.error('Full error object:', error);
-      });
-  };
-
-  
-  const handleSubmit = () => {
-    const patientData = {
-      id : Math.floor(Math.random() * 200),
-      first_name: firstName,
-      last_name:lastName,
-      device_id: Math.floor(Math.random() * 200),
-      dob,
-      gender,
-    };
-
-    console.log('Submitting patientData:', patientData);
-    setFirstName('');
-    setLastName('');
-    setDOB('');
-    setGender('');
-
-    axios.post('http://localhost:8000/api/user', patientData)
-    .then(response => {
-      // Handle success (e.g., show a success message, navigate to a new screen)
-      console.log('Patient registered:', response.data);
-      alert('Patient registered successfully! Device ID:' + response.data.device_id)})
-    .catch(error => {
-      // Handle error (e.g., show an error message)
-      alert('Error registering patient. Please try again.');
-      if (error.response) {
-        // The request was made and the server responded with a status code
-        // that falls out of the range of 2xx
-        console.error('Server responded with non-2xx status:', error.response.status);
-        console.error('Response data:', error.response.data);
-        console.error('Response headers:', error.response.headers);
-      } else if (error.request) {
-        // The request was made but no response was received
-        console.error('No response received from the server');
-        console.error('Request data:', error.request);
-      } else {
-        // Something happened in setting up the request that triggered an Error
-        console.error('Error setting up the request:', error.message);
-      }
-      console.error('Full error object:', error);
-    });
-  
-  };
-
-  const formatDOB = (input) => {
-    if (input.length <= 10) {
-      // Remove any non-numeric characters
-      const numericInput = input.replace(/\D/g, '');
-
-      if (numericInput.length <= 2) {
-        // Format as MM
-        return numericInput;
-      } else if (numericInput.length <= 4) {
-        // Format as MM/DD
-        return `${numericInput.slice(0, 2)}/${numericInput.slice(2)}`;
-      } else {
-        // Format as MM/DD/YYYY
-        return `${numericInput.slice(0, 2)}/${numericInput.slice(2, 4)}/${numericInput.slice(4, 8)}`;
-      }
-    }
-    return input.slice(0, 10);
-  };
-  
-  const columns = useMemo(
-    () => [
-      {
-        accessorKey: 'id', //access nested data with dot notation
-        header: 'Patient ID',
-        size: 3,
-      },
-      {
-        accessorKey: 'device_id', //access nested data with dot notation
-        header: 'Device ID',
-        size: 3,
-      },
-      {
-        accessorKey: 'first_name', //access nested data with dot notation
-        header: 'First Name',
-        size: 20,
-      },
-      {
-        accessorKey: 'last_name', //access nested data with dot notation
-        header: 'Last Name',
-        size: 20,
-      },
-      {
-        accessorKey: 'dob', //access nested data with dot notation
-        header: 'DOB',
-        size: 10,
-      },
-      {
-        accessorKey: 'gender', //access nested data with dot notation
-        header: 'Gender',
-        size: 10,
-      },
-    ],
-    [],
-  );
-
-
-  const table = useMaterialReactTable({
-    columns,
-    data: allData,
-    initialState: {density: 'compact'},
-    muiTableHeadCellProps: {
-      style: {
-        backgroundColor: '#6373ff',
-        color: 'white',
-      },
-    },
-    muiTableFooterCellProps: {
-      style: {
-        backgroundColor: '#6373ff',
-        color: 'white',
-      },
-    },
-    
-  });
 
   return (
     <View style={styles.container}>
@@ -245,80 +37,17 @@ export default function App() {
           }
             containerStyle={styles.headerContainer}
       />
-      <form style={styles.form}>
-        <Text style={styles.header2}>Register a Patient</Text>
-        <Input
-          style={styles.input}
-          placeholder="First Name"
-          value={firstName}
-          onChangeText={text => setFirstName(text)}
-        />
-        <Input 
-          style={styles.input}
-          placeholder="Last Name" 
-          value={lastName}
-          onChangeText={text => setLastName(text)}
-        />
-        <Input 
-          style={styles.input}
-          placeholder="DOB: (MM/DD/YYYY)"
-          value={formatDOB(dob)}
-          onChangeText={text => setDOB(formatDOB(text))}
-        />
-        <Input 
-          style={styles.input}
-          placeholder="Gender"
-          value={gender}
-          onChangeText={text => setGender(text)}
-        />
-        <Button type="submit" style={styles.submit} title="Submit" titleStyle={{ color: 'white'}} onPress={handleSubmit}></Button>
-      </form>
-      <form style={styles.form}>
-        <Text style={styles.header2}>Find a Patient</Text>
-        <Input
-          style={styles.input}  
-          placeholder="Device ID"
-          value={deviceId}
-          onChangeText={text => {
-            // Allow editing only if the Device ID field is empty
-            if (!searchLast) {
-              setDeviceId(text);
-            }
-          }}
-        />
-        <Text style={styles.header2}>OR</Text>
-        <Input
-          style={styles.input}
-          placeholder="Last Name"
-          value={searchLast}
-          onChangeText={text => {
-            // Allow editing only if the Device ID field is empty
-            if (!deviceId) {
-              setSearchLast(text);
-            }
-          }}
-        />
-        <Button type="submit" style={styles.search} onPress={handleSearch} ><SearchOutlined /></Button>
-        {/* <Text style={styles.paragraph}>First Name: {searchResults.first_name}</Text> */}
-        <FlatList
-          data={[searchResults.id, searchResults.first_name, searchResults.last_name, searchResults.dob, searchResults.gender]}
-          renderItem={({ item }) => ( 
-            <Text style={styles.paragraph}>
-              {item}
-            </Text>
-          )}
-        />
-      </form>
+      <div style={{ display: 'flex', justifyContent: 'center' }}>
+        <Register />
+        <Search />
+      </div>
       <form style={styles.form}>
         <Text style={styles.header2}>All Data</Text>
         <br></br>
         <br></br>
-          <MaterialReactTable 
-            table={table}
-          />
+          <AllDataTable/>
         <br></br>
       </form>
-      
       </SafeAreaProvider>
       <StatusBar style="auto" />
     </View>
@@ -333,26 +62,13 @@ const styles = StyleSheet.create({
     backgroundColor: '#19173D',
     borderWidth: .3,
   },
-  input: {
-    color: '#FFFFFF',
-  },
   form: { 
     alignSelf: 'center',
     padding: 10,
     margin: 10,
-    width: windowWidth/2,
+    width: '60%',
     backgroundColor: '#292847',
     borderRadius: 10,
-  },
-  submit: {
-    backgroundColor: '#5A6BFF',
-    color: '#FFFFFF',
-    margin: 10,
-  },
-  search: {
-    backgroundColor: '#5A6BFF',
-    color: '#FFFFFF',
-    padding: 5,
   },
   container: {
     backgroundColor: '#19173D',
